requirements.txt
fastapi==0.110.0 
uvicorn==0.29.0
pytest==5.3.2
requests>=2.32.0
python-dotenv==1.0.1
websockets<12.0
gradio-client==0.10.1
ollama==0.1.9
langchain
langchain-community
chromadb
pypdf
wget